+++ quilt/changelog.md
@@ -9,3 +9,4 @@
 - Renamed microphone amplification to microphone gain
 - Fixed French translation
 - Added audio level meters to adjust volumes section
+- Reworked player volume adjustment